Step 12: Change PlatformWord from usize to cfg-based u64/u32
Remove the compile_error! gate for non-64-bit targets. PlatformWord is
now u64 on 64-bit and u32 on 32-bit. All inventory/spill arrays,
SmallCounters, and SelectSmall use PlatformWord with explicit casts at
read/write boundaries. Rank9/Select9 remain pinned to u64.

Dense PartialArray now uses Box<[u64]> backing (Rank9 requires u64).
SelectSmall/SelectZeroSmall SUPERBLOCK_BIT_SIZE uses cfg to avoid
1usize << 32 overflow on 32-bit. thread-priority made optional behind
the rayon feature for wasm portability.

Cross-compiles successfully for wasm32-wasip1.

+++ tests/test_bit_field_vec.rs
@@ -663,7 +663,7 @@ fn test_iter_from() {
     for i in 0..b.len() {
         let mut iter = b.iter_from(i);
         for j in i..b.len() {
-            assert_eq!(iter.next(), Some(j));
+            assert_eq!(iter.next(), Some(j as u64));
         }
     }
 }
@@ -675,7 +675,7 @@ fn test_iter_double_ended() {
     // next_back only
     let mut iter = b.into_iter();
     for j in (0..b.len()).rev() {
-        assert_eq!(iter.next_back(), Some(j));
+        assert_eq!(iter.next_back(), Some(j as u64));
     }
     assert_eq!(iter.next_back(), None);
